separate pending state update verifier
