+++ packages/plasma-light-client/src/LightClient.ts
@@ -10,18 +10,16 @@ import {
   CompiledPredicate,
   DeciderManager,
   DeciderConfig,
-  Challenge,
-  hint as Hint
+  Challenge
 } from '@cryptoeconomicslab/ovm'
 import {
   Address,
   Bytes,
-  FixedBytes,
   BigNumber,
   Property,
   Range
 } from '@cryptoeconomicslab/primitives'
-import { KeyValueStore, putWitness } from '@cryptoeconomicslab/db'
+import { KeyValueStore } from '@cryptoeconomicslab/db'
 import {
   ICommitmentContract,
   IDepositContract,
@@ -31,17 +29,8 @@ import {
 } from '@cryptoeconomicslab/contract'
 import { Wallet } from '@cryptoeconomicslab/wallet'
 import { decodeStructable } from '@cryptoeconomicslab/coder'
-import {
-  DoubleLayerInclusionProof,
-  DoubleLayerTreeVerifier,
-  DoubleLayerTreeLeaf
-} from '@cryptoeconomicslab/merkle-tree'
-import { Keccak256 } from '@cryptoeconomicslab/hash'
 import JSBI from 'jsbi'
-import UserAction, {
-  createDepositUserAction,
-  createSendUserAction
-} from './UserAction'
+import UserAction, { createDepositUserAction } from './UserAction'
 import EventEmitter from 'event-emitter'
 import {
   StateUpdateRepository,
@@ -52,6 +41,7 @@ import {
 } from './repository'
 import { StateSyncer } from './usecase/StateSyncer'
 import { ExitUsecase } from './usecase/ExitUsecase'
+import { PendingStateUpdatesVerifier } from './verifier/PendingStateUpdatesVerifier'
 import APIClient from './APIClient'
 import getTokenManager, { TokenManager } from './managers/TokenManager'
 import { executeChallenge } from './helper/challenge'
@@ -80,6 +70,7 @@ export default class LightClient {
   private tokenManager: TokenManager
   private stateSyncer: StateSyncer
   private exitUsecase: ExitUsecase
+  private pendingStateUpdatesVerifier: PendingStateUpdatesVerifier
 
   constructor(
     private wallet: Wallet,
@@ -118,6 +109,11 @@ export default class LightClient {
       this.commitmentContract,
       this.ownershipPayoutContract,
       this.deciderManager
+    )
+    this.pendingStateUpdatesVerifier = new PendingStateUpdatesVerifier(
+      this.ee,
+      this.witnessDb,
+      this.apiClient
     )
   }
 
@@ -207,7 +203,7 @@ export default class LightClient {
       async (blockNumber, root) => {
         console.log('new block submitted event:', root.toHexString())
         await this.stateSyncer.sync(blockNumber, Address.from(this.address))
-        await this.verifyPendingStateUpdates(blockNumber)
+        await this.pendingStateUpdatesVerifier.verify(blockNumber)
       }
     )
     this.commitmentContract.startWatchingEvents()
@@ -229,101 +225,6 @@ export default class LightClient {
         depositContract.unsubscribeAll()
       }
     })
-  }
-
-  /**
-   * checks if pending state updates which basically are state updates client transfered,
-   *  have been included in the block.
-   * @param blockNumber block number to verify pending state updates
-   */
-  private async verifyPendingStateUpdates(blockNumber: BigNumber) {
-    console.group('VERIFY PENDING STATE UPDATES: ', blockNumber.raw)
-    const stateUpdateRepository = await StateUpdateRepository.init(
-      this.witnessDb
-    )
-
-    this.tokenManager.depositContractAddresses.forEach(async addr => {
-      const pendingStateUpdates = await stateUpdateRepository.getPendingStateUpdates(
-        addr,
-        new Range(BigNumber.from(0), BigNumber.MAX_NUMBER)
-      )
-      const verifier = new DoubleLayerTreeVerifier()
-      const syncRepository = await SyncRepository.init(this.witnessDb)
-      const root = await syncRepository.getBlockRoot(blockNumber)
-      if (!root) {
-        return
-      }
-
-      pendingStateUpdates.forEach(async su => {
-        console.info(
-          `Verify pended state update: (${su.range.start.data.toString()}, ${su.range.end.data.toString()})`
-        )
-        let res
-        try {
-          res = await this.apiClient.inclusionProof(su)
-        } catch (e) {
-          return
-        }
-        const { coder } = ovmContext
-        const inclusionProof = decodeStructable(
-          DoubleLayerInclusionProof,
-          coder,
-          Bytes.fromHexString(res.data.data)
-        )
-        const leaf = new DoubleLayerTreeLeaf(
-          su.depositContractAddress,
-          su.range.start,
-          FixedBytes.from(
-            32,
-            Keccak256.hash(coder.encode(su.property.toStruct())).data
-          )
-        )
-        if (verifier.verifyInclusion(leaf, su.range, root, inclusionProof)) {
-          console.info(
-            `Pended state update (${su.range.start.data.toString()}, ${su.range.end.data.toString()}) verified. remove from stateDB`
-          )
-          await stateUpdateRepository.removePendingStateUpdate(
-            su.depositContractAddress,
-            su.range
-          )
-
-          // store inclusionProof as witness
-          const hint = Hint.createInclusionProofHint(
-            blockNumber,
-            su.depositContractAddress,
-            su.range
-          )
-          await putWitness(
-            this.witnessDb,
-            hint,
-            Bytes.fromHexString(res.data.data)
-          )
-
-          // store send user action
-          const { range } = su
-          const owner = getOwner(su)
-          const tokenContractAddress = this.tokenManager.getTokenContractAddress(
-            su.depositContractAddress
-          )
-          if (!tokenContractAddress)
-            throw new Error('Token Contract Address not found')
-          const actionRepository = await UserActionRepository.init(
-            this.witnessDb
-          )
-          const action = createSendUserAction(
-            Address.from(tokenContractAddress),
-            range,
-            owner,
-            su.blockNumber
-          )
-          await actionRepository.insertAction(su.blockNumber, range, action)
-
-          this.ee.emit(UserActionEvent.SEND, action)
-          this.ee.emit(EmitterEvent.TRANSFER_COMPLETE, su)
-        }
-      })
-    })
-    console.groupEnd()
   }
 
   /**
